Updated for use with my particular server
Looking to see if i can get this to work on my own server by changing the config output to the particular folder location on my server. Will chaining  this one line work in updating it through the shell script? I'm a bit of a noob and kinda just learning this stuff.

Thanks!